+++ lib/src/common/field/suggestions_field.dart
@@ -5,11 +5,11 @@ import 'package:flutter_typeahead/src/common/box/suggestions_box.dart';
 import 'package:flutter_typeahead/src/common/base/suggestions_controller.dart';
 import 'package:flutter_typeahead/src/common/base/types.dart';
 import 'package:flutter_typeahead/src/common/field/suggestions_field_focus_connector.dart';
+import 'package:flutter_typeahead/src/common/field/suggestions_field_highlight_connector.dart';
 import 'package:flutter_typeahead/src/common/field/suggestions_field_keyboard_connector.dart';
 import 'package:flutter_typeahead/src/common/field/suggestions_field_box_connector.dart';
 import 'package:flutter_typeahead/src/common/field/suggestions_field_select_connector.dart';
 import 'package:flutter_typeahead/src/common/field/suggestions_field_tap_connector.dart';
-import 'package:flutter_typeahead/src/common/field/suggestions_field_traversal_connector.dart';
 
 /// A widget that displays a list of suggestions above or below another widget.
 class SuggestionsField<T> extends StatefulWidget {
@@ -23,10 +23,12 @@ class SuggestionsField<T> extends StatefulWidget {
     this.direction,
     this.autoFlipDirection = false,
     this.autoFlipMinHeight = 64,
+    this.showOnFocus = true,
     this.hideOnUnfocus = true,
     this.hideOnSelect = true,
     this.hideWithKeyboard = true,
     this.constraints,
+    this.constrainWidth = true,
     this.offset,
     this.scrollController,
     this.decorationBuilder,
@@ -78,8 +80,16 @@ class SuggestionsField<T> extends StatefulWidget {
   /// {@endtemplate}
   final BoxConstraints? constraints;
 
+  /// {@template flutter_typeahead.SuggestionsField.constrainWidth}
+  /// Whether the suggestions box should be constrained to the width of the child.
+  ///
+  /// Defaults to true.
+  /// {@endtemplate}
+  final bool constrainWidth;
+
   /// {@template flutter_typeahead.SuggestionsField.offset}
   /// The offset of the suggestions box.
+  /// The x value of this property will be applied symmetrically.
   /// The value is automatically flipped if the suggestions box is flipped.
   ///
   /// Defaults to `Offset(0, 5)`.
@@ -99,11 +109,20 @@ class SuggestionsField<T> extends StatefulWidget {
   final bool autoFlipDirection;
 
   /// {@template flutter_typeahead.SuggestionsField.autoFlipMinHeight}
-  /// The minimum height the suggesttions box can have before attempting to flip.
+  /// The minimum height the suggestions box can have before attempting to flip.
   ///
   /// Defaults to 64.
   /// {@endtemplate}
   final double autoFlipMinHeight;
+
+  /// {@template flutter_typeahead.SuggestionsField.showOnFocus}
+  /// Whether the suggestions box should be shown when the child of the suggestions box gains focus.
+  ///
+  /// If disabled, the suggestions box will remain closed when the user taps on the child of the suggestions box.
+  ///
+  /// Defaults to true.
+  /// {@endtemplate}
+  final bool showOnFocus;
 
   /// {@template flutter_typeahead.SuggestionsField.hideOnUnfocus}
   /// Whether the suggestions box should be hidden when the child of the suggestions box loses focus.
@@ -204,8 +223,16 @@ class _SuggestionsFieldState<T> extends State<SuggestionsField<T>> {
           VerticalDirection.up => AxisDirection.up,
           VerticalDirection.down => AxisDirection.down,
         },
-        offset: widget.offset ?? const Offset(0, 5),
-        followHeight: false,
+        padding: EdgeInsets.only(
+          top: widget.offset?.dy ?? 5,
+          left: widget.offset?.dx ?? 0,
+          right: widget.offset?.dx ?? 0,
+        ),
+        anchor: FloaterAnchor.only(
+          left: widget.constrainWidth,
+          right: widget.constrainWidth,
+          bottom: false,
+        ),
         autoFlip: widget.autoFlipDirection,
         autoFlipHeight: widget.autoFlipMinHeight,
         builder: (context) {
@@ -263,16 +290,16 @@ class _SuggestionsFieldState<T> extends State<SuggestionsField<T>> {
           link: link,
           child: ConnectorWidget(
             value: controller,
-            connect: (value) => value.resizes.listen((_) => onResize()),
+            connect: (value) => value.$resizes.listen((_) => onResize()),
             disconnect: (value, key) => key?.cancel(),
             child: SuggestionsFieldFocusConnector<T>(
               controller: controller,
               focusNode: widget.focusNode,
-              child: SuggestionsFieldTraversalConnector<T>(
+              child: SuggestionsFieldHighlightConnector<T>(
                 controller: controller,
-                focusNode: widget.focusNode,
                 child: SuggestionsFieldBoxConnector<T>(
                   controller: controller,
+                  showOnFocus: widget.showOnFocus,
                   hideOnUnfocus: widget.hideOnUnfocus,
                   child: SuggestionsFieldKeyboardConnector<T>(
                     controller: controller,