+++ lib/src/common/field/typeahead_field.dart
@@ -28,6 +28,7 @@ abstract class RawTypeAheadField<T> extends StatefulWidget {
     this.hideOnEmpty = false,
     this.hideOnError = false,
     this.hideOnLoading = false,
+    this.showOnFocus = true,
     this.hideOnUnfocus = true,
     this.hideWithKeyboard = true,
     this.hideOnSelect = true,
@@ -44,6 +45,7 @@ abstract class RawTypeAheadField<T> extends StatefulWidget {
     this.decorationBuilder,
     this.listBuilder,
     this.constraints,
+    this.constrainWidth = true,
     this.offset,
     this.icon,
   });
@@ -74,6 +76,9 @@ abstract class RawTypeAheadField<T> extends StatefulWidget {
   /// {@macro flutter_typeahead.SuggestionsField.constraints}
   final BoxConstraints? constraints;
 
+  /// {@macro flutter_typeahead.SuggestionsField.constrainWidth}
+  final bool constrainWidth;
+
   /// {@macro flutter_typeahead.SuggestionsField.offset}
   final Offset? offset;
 
@@ -82,6 +87,9 @@ abstract class RawTypeAheadField<T> extends StatefulWidget {
 
   /// {@macro flutter_typeahead.SuggestionsField.autoFlipMinHeight}
   final double autoFlipMinHeight;
+
+  /// {@macro flutter_typeahead.SuggestionsField.showOnFocus}
+  final bool showOnFocus;
 
   /// {@macro flutter_typeahead.SuggestionsField.hideOnUnfocus}
   final bool hideOnUnfocus;
@@ -195,10 +203,12 @@ class _RawTypeAheadFieldState<T> extends State<RawTypeAheadField<T>> {
       direction: widget.direction,
       autoFlipDirection: widget.autoFlipDirection,
       autoFlipMinHeight: widget.autoFlipMinHeight,
+      showOnFocus: widget.showOnFocus,
       hideOnUnfocus: widget.hideOnUnfocus,
       hideOnSelect: widget.hideOnSelect,
       hideWithKeyboard: widget.hideWithKeyboard,
       constraints: widget.constraints,
+      constrainWidth: widget.constrainWidth,
       offset: widget.offset,
       scrollController: widget.scrollController,
       decorationBuilder: (context, child) => TextFieldTapRegion(