+++ lib/src/cupertino/typeahead_field.dart
@@ -25,6 +25,7 @@ class CupertinoTypeAheadField<T> extends RawTypeAheadField<T> {
     super.hideOnEmpty,
     super.hideOnError,
     super.hideOnLoading,
+    super.showOnFocus,
     super.hideOnUnfocus,
     super.hideWithKeyboard,
     super.hideOnSelect,
@@ -41,6 +42,7 @@ class CupertinoTypeAheadField<T> extends RawTypeAheadField<T> {
     DecorationBuilder? decorationBuilder,
     super.listBuilder,
     super.constraints,
+    super.constrainWidth,
     super.offset,
   }) : super(
           builder: builder ?? TypeAheadCupertinoDefaults.builder,